--- components/Calculator.js
@@ -1,44 +1,0 @@
-import React, { Component } from 'react';
-
-class Calculator extends Component {
-  constructor(props) {
-      super(props);
-      this.state = {};
-  }
-
-  render() {
-      return(
-          <div class="Calculator">
-              <input value='0' />
-              <button type='submit' className='calculatorBtn'>AC</button>
-              <button type='submit' className='calculatorBtn'></button>
-              <button type='submit' className='calculatorBtn'></button>
-              <button type='submit' className='calculatorBtn'></button>
-              <button type='submit' className='calculatorBtn'></button>
-              <button type='submit' className='calculatorBtn'></button>
-              <button type='submit' className='calculatorBtn'></button>
-              <button type='submit' className='calculatorBtn'></button>
-              <button type='submit' className='calculatorBtn'></button>
-              <button type='submit' className='calculatorBtn'></button>
-              <button type='submit' className='calculatorBtn'></button>
-              <button type='submit' className='calculatorBtn'></button>
-              <button type='submit' className='calculatorBtn'></button>
-              <button type='submit' className='calculatorBtn'></button>
-          </div>
-      )
-  }
-} 
-
-
-
-
-
-
-
-
-
-
-
-
-export default expression;
-
