+++ src/components/Calculator.js
@@ -1,0 +1,30 @@
+import React, { Component } from 'react';
+
+class Calculator extends Component {
+  constructor(props) {
+      super(props);
+      this.state = {};
+  }
+
+  render() {
+      return(
+          <div class="Calculator">
+             <h1>Hello world</h1>>
+          </div>
+      )
+  }
+} 
+
+
+
+
+
+
+
+
+
+
+
+
+export default Calculator;;
+
