+++ src/index.js
@@ -3,6 +3,8 @@ import ReactDOM from 'react-dom';
 import './index.css';
 import App from './App';
 import reportWebVitals from './reportWebVitals';
+import Calculator from './components'
+
 
 ReactDOM.render(
   <React.StrictMode>
